requirements.txt
Django>=2.1.1
django-bootstrap4==0.0.6
django-crispy-forms==1.7.2
requests==2.18.4
django-braces==1.13.0
